requirements.txt
aiohttp~=3.9.1
aiohttp-jinja2==1.6
Jinja2==3.1.2
aiopg~=1.4.0
alembic~=1.9.2
aiobotocore==2.9.0
aiojobs~=1.1.0
aiohttp_session
redis>=4.2.0rc1
psycopg2-binary
SQLAlchemy~=1.4.46
pyjwt
bcrypt
uvloop
pytz==2022.7.1
python-dotenv~=0.21.0
cerberus
gunicorn==20.1.0
aiohttp_cors
pika
aio-pika==9.3.1
aiormq==6.7.7
pymongo==4.6.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
XlsxWriter==3.1.9
aiosmtplib==3.0.1
